% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cols.R
\name{add_bmi}
\alias{add_bmi}
\title{Add BMI column}
\usage{
add_bmi(df)
}
\arguments{
\item{df}{a data frame containing a \code{height} and \code{weight} column.}
}
\value{
a data frame
}
\description{
Calculates a BMI from \code{height} and \code{weight} and binds an additional column to the supplied data frame.
}
\examples{
df <- data.frame(height = c(70, 72, 64), weight = c(150, 160, 170))
df <- df \%>\% add_bmi()

}
