% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cols.R
\name{add_dindo}
\alias{add_dindo}
\title{Add Clavien-Dindo classification column}
\usage{
add_dindo(df)
}
\arguments{
\item{df}{a data frame containing any post-operative complication columns.}
}
\value{
a data frame
}
\description{
Determines a Clavien-Dindo classification from post-operative complication columns and binds an additional column to the supplied data frame.
}
\examples{
supinfec <- c(TRUE, FALSE, FALSE)
returnor <- c(FALSE, TRUE, FALSE)
othseshock <- c(FALSE, FALSE, TRUE)
df <- data.frame(supinfec, returnor, othseshock)

}
