% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cols.R
\name{add_dindo_group}
\alias{add_dindo_group}
\title{Add a Dindo-Clavien Classification Group}
\usage{
add_dindo_group(df)
}
\arguments{
\item{df}{a data frame containing any post-operative complication columns.}
}
\description{
Determines a Clavien-Dindo classification group from post-operative complication columns and binds an additional column to the supplied data frame.
}
\details{
Classifies Dindo-Clavien classification groups 1 and 2 as "minor" and groups 3 and above as "major".
}
\examples{
supinfec <- c(TRUE, FALSE, FALSE)
returnor <- c(FALSE, TRUE, FALSE)
othseshock <- c(FALSE, FALSE, TRUE)
df <- data.frame(supinfec, returnor, othseshock)
df <- df \%>\% add_dindo_group()

}
