% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cols.R
\name{add_postopindependent}
\alias{add_postopindependent}
\title{Add postoperatively independent column}
\usage{
add_postopindependent(df)
}
\arguments{
\item{df}{a data frame containing a \code{fnstatus1} column.}
}
\value{
a data frame
}
\description{
Creates a column that indicates if a patient is post-operatively functionally independent or not.
}
\examples{
df <- data.frame(fnstatus1 = c(1, 2, 3))
df <- df \%>\% add_postopindependent()

}
