% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cols.R
\name{add_preopindependent}
\alias{add_preopindependent}
\title{Add preoperatively independent column}
\usage{
add_preopindependent(df)
}
\arguments{
\item{df}{a data frame containing a \code{fnstatus2} column.}
}
\value{
a data frame
}
\description{
Creates a column that indicates if a patient is preoperatively functionally independent or not.
}
\examples{
df <- data.frame(fnstatus2 = c(1, 2, 3))
df <- df \%>\% add_preopindependent()

}
