% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{assign_label}
\alias{assign_label}
\title{Assign a label to a column}
\usage{
assign_label(df, col)
}
\arguments{
\item{df}{a data frame}

\item{col}{a column}
}
\value{
A dataframe with the new labels
}
\description{
Assigns a label to a data frame column using \code{Hmisc::\link[Hmisc:label]{label()}}
}
\details{
Fetches label from \code{nsqipBileSpill::\link[nsqipBileSpill:fetch_label]{fetch_label()}}.

In order to ascribe your own labels, you must update the \code{nsqipBileSpill::\link[nsqipBileSpill:fetch_label]{fetch_label()}} function.
}
\examples{
df <- data.frame(age = c(1,2,3))
df \%>\% assign_label(age)

}
