% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{assign_labels}
\alias{assign_labels}
\title{Assign labels to a data frame}
\usage{
assign_labels(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
A dataframe with the new labels
}
\description{
Assigns labels to data frame columns using \code{Hmisc::\link[Hmisc:label]{label()}}
}
\details{
Fetches labels from \code{nsqipBileSpill::\link[nsqipBileSpill:fetch_label]{fetch_label()}}.

In order to ascribe your own labels, you must update the \code{nsqipBileSpill::\link[nsqipBileSpill:fetch_label]{fetch_label()}} function.
}
\examples{
df <- data.frame(age = c(25, 32, 47), height = c(72, 65, 70), weight = c(180, 210, 190))
df \%>\% assign_labels()

}
