% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avg}
\alias{avg}
\title{Calculate an average}
\usage{
avg(vec)
}
\arguments{
\item{vec}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Returns an average from a numeric vector, automatically removes all NA values.
}
\examples{
avg(c(10, 10, 10, NA))

}
