% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparison}
\alias{comparison}
\title{Compare means/proportions across a group}
\usage{
comparison(df, group, ...)
}
\arguments{
\item{df}{a dataframe}

\item{group}{the grouping variable; must be a \code{logical} vector}

\item{...}{columns to compare means/proportions}
}
\value{
a data frame
}
\description{
Compares the means/proportions across grouping by \code{group}. Produces a p- value for each comparsion.
}
\details{
This function produces either a \code{t.test} or a \code{chisq.test} on each of the columns in \code{...}. If the variable \code{is.numeric}, conducts \code{t.test}. If the variable \code{is.logical} or \code{is.factor}, conducts \code{chisq.test}. Note that if you are piping into this function with a variable created by \code{dplyr::group_by()} as as your grouping function, you must first \code{dplyr::ungroup()} prior to piping into \code{compare()} (see example).
}
\examples{
mtcars \%>\% dplyr::group_by(hp > 150) \%>\% dplyr::ungroup() \%>\% comparison(`hp > 150`, hp, am)

}
