% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{fetch_label}
\alias{fetch_label}
\title{Fetch a label}
\usage{
fetch_label(col)
}
\arguments{
\item{col}{a string representing a column name from a dataframe}
}
\value{
a character vector
}
\description{
Fetches a character vector to be used as a label for a column from a pre-determined list of labels.
}
\examples{
fetch_label("age")

}
