% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_percent}
\alias{get_percent}
\title{Calculate percentage}
\usage{
get_percent(vec)
}
\arguments{
\item{vec}{a logical vector}
}
\value{
a numeric vector
}
\description{
Returns a percentage TRUE given a logical vector.
}
\examples{
get_percent(mtcars$hp > 150)

}
