% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{make_formula}
\alias{make_formula}
\title{Create a formula}
\usage{
make_formula(..., lhs = TRUE, op = "+", group = NULL)
}
\arguments{
\item{...}{any number of arguments to compose the formula}

\item{lhs}{a boolean indicating if the formula has a left hand side of the argument}

\item{op}{the operand acting upon the arguments of the right side of the formula.}

\item{group}{an argument to use as a grouping variable to facet by}
}
\value{
a formula
}
\description{
Creates a new formula object to be used anywhere formulas are used (i.e, \code{glm}).
}
\details{
If \code{lhs} is \code{TRUE}, the first argument provided is used as the left hand side of the formula.
The \code{group} paramenter will add \verb{| group} to the end of the formula. This is useful for packages that support faceting by grouping variables for the purposes of tables or graphs.
}
\examples{
make_formula(var1, var2, var3)
make_formula(var1, var2, var3, lhs = FALSE)
make_formula(var1, var2, var3, lhs = FALSE, group = var4)

}
