% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_table1}
\alias{make_table1}
\title{Make a Table 1}
\usage{
make_table1(df, group, ..., negative_label = "No", caption, table_num)
}
\arguments{
\item{df}{a dataframe}

\item{group}{a group to split the columns by. Must be a logical or a factor with two levels.}

\item{...}{a list of columns to be included in the table}

\item{negative_label}{a word to prepend to the label of the negative \code{group} (see example)}

\item{caption}{an optional caption}

\item{table_num}{an optional table number}
}
\value{
an HTML table
}
\description{
This function provides a wrapper around \code{table1::\link[table1:table1]{table1}} that allows it to
create a column for P-values and perform the appropriate statistical tests. It also adds functionality
for captions and table numbering.
}
\details{
Currently only works if you intend to split by the variable \code{group}, which must be a logical or a factor with two levels.
}
\examples{
diamonds <- ggplot2::diamonds
diamonds$expensive <- diamonds$price > 500
Hmisc::label(diamonds$depth) <- "Depth"
Hmisc::label(diamonds$table) <- "Table"
Hmisc::label(diamonds$price) <- "Price"
Hmisc::label(diamonds$clarity) <- "Clarity"
Hmisc::label(diamonds$cut) <- "Cut"
Hmisc::label(diamonds$expensive) <- "Expensive"
diamonds \%>\% make_table1(group = expensive, depth, clarity, cut, negative_label = "Not")

}
