% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-notin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in operator}
\usage{
lhs \%nin\% rhs
}
\arguments{
\item{lhs}{a vector containing values to be matched}

\item{rhs}{a vector to compare \code{lhs} to}
}
\description{
Negates the \code{\%in\%} operator. See \code{base::\link[base:match]{match}} for details.
}
\examples{
"dog" \%nin\% c("cat","giraffe","elephant")
"dog" \%nin\% c("dog","giraffe","elephant")

}
