% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Calculate an odds ratio}
\usage{
odds_ratio(df, exposure, outcome)
}
\arguments{
\item{df}{a data frame}

\item{exposure}{a logical vector representing a risk factor or exposure}

\item{outcome}{a logical vector representing an outcome}
}
\value{
a list object produced by \code{epitools::\link[epitools:oddsratio]{oddsratio()}}
}
\description{
Given a data frame, a risk factor column, and an outcome column, calculate odds ratio via \code{epitools::\link[epitools:oddsratio]{oddsratio()}}. Allows non-standard evaluation for use in \code{dplyr} pipes.
}
\examples{
df <- data.frame(poison = c(TRUE, TRUE, FALSE, FALSE), death = c(TRUE, FALSE, TRUE, FALSE))
df \%>\% odds_ratio(poison, death)

}
