% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{an input}

\item{rhs}{function to apply to input from \code{lhs}}
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
