% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cols.R
\name{pull_cell}
\alias{pull_cell}
\title{Select a cell from a dataframe given a row and a column}
\usage{
pull_cell(df, rowname, col, percent = FALSE)
}
\arguments{
\item{df}{a dataframe}

\item{rowname}{the name (or id) of a row}

\item{col}{the name of a column}

\item{percent}{return value multiplied by 100}
}
\value{
a single cell from a data frame that is of the class of the value's containing column
}
\description{
Given a row (identified by row names or a character column called "rowname"), select the cell that intersects with a given column.
}
\details{
This function requires a data frame either have row names, a column called "rowname", or that a row be referred to by its ID if neither of the previous conditions are true.
\code{percent} is useful for returning numeric values reported as rates that you would like reported as percentages.
}
\examples{
pull_cell(mtcars, `Mazda RX4`, mpg, percent = FALSE)

}
