% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cols.R
\name{remove_missing}
\alias{remove_missing}
\title{Remove records with missing values}
\usage{
remove_missing(df, ...)
}
\arguments{
\item{df}{a dataframe}

\item{...}{columns to check for missing values}
}
\value{
a dataframe
}
\description{
Removes all records from the specified columns that have missing values via filtering for records by \code{!is.na()}.
}
\examples{
df <- data.frame(a = c(1, 2, NA), b = c(3, NA, 4), c = c(NA, NA, NA))
df \%>\% remove_missing(a, b)

}
